+++ Sources/StreamChat/Database/DTOs/ChannelDTO.swift
@@ -1,5 +1,5 @@
 //
-// Copyright © 2021 Stream.io Inc. All rights reserved.
+// Copyright © 2022 Stream.io Inc. All rights reserved.
 //
 
 import CoreData
@@ -61,6 +61,7 @@ class ChannelDTO: NSManagedObject {
     @NSManaged var reads: Set<ChannelReadDTO>
     @NSManaged var attachments: Set<AttachmentDTO>
     @NSManaged var watchers: Set<UserDTO>
+    @NSManaged var memberListQueries: Set<ChannelMemberListQueryDTO>
 
     override func willSave() {
         super.willSave()