+++ Sources/StreamChatUI/ChatChannel/ChatChannelVC_Tests.swift
@@ -1,5 +1,5 @@
 //
-// Copyright © 2021 Stream.io Inc. All rights reserved.
+// Copyright © 2022 Stream.io Inc. All rights reserved.
 //
 
 import StreamChat
@@ -206,6 +206,7 @@ final class ChatChannelVC_Tests: XCTestCase {
         components.messageListView = TestMessageListView.self
         components.messageComposerView = TestComposerView.self
         vc.components = components
+        vc.messageListVC.components = components
         
         // Simulate view loading
         _ = vc.view